ParsedUpdate to capture deserialization errors (#144)
`ParsedUpdate` type was introduce in order to be able to represent parsing issues in the GetUpdates endpoint.
With the previous implementation, when a new or unknown type of message was received by the client, the parser would fail and cause the bot to stop parsing new messages.
The new implementation extracts the update_id message in case of parsing issue in order to acknowledge the message and receive the next ones. It also extracts and logs the decoding error from Circe in order to fix them in the next release of the library.

This also includes a fix to run some tests that were skipped so far as well as correctly serializing the CallbackGame case object.