[api] Add missing fields to Chat and Message (#192)
Some fields were missing from `Message` and `Chat` class.
Those were added. Those two classes are close to complete, but some fields are still missing because they would require the introduction of a new case class for which I don't have the time right now (and this is non-blocking/breaking).